+++ npm.go
@@ -18,6 +18,36 @@ type PackageJSON struct {
 	OptionalDependencies map[string]string `json:"optionalDependencies,omitempty"`
 }
 
+type NpmResponse struct {
+	ID   string `json:"_id"`
+	Name string `json:"name"`
+	Time struct {
+		Unpublished NpmResponseUnpublished `json:"unpublished"`
+	} `json:"time"`
+}
+
+type NpmResponseUnpublished struct {
+		Maintainers []struct {
+			Email string `json:"email"`
+			Name  string `json:"name"`
+		} `json:"maintainers"`
+		Name string `json:"name"`
+		Tags struct {
+			Latest string `json:"latest"`
+		} `json:"tags"`
+		Time     time.Time `json:"time"`
+		Versions []string  `json:"versions"`
+}
+
+// NotAvailable returns true if the package has its all versions unpublished making it susceptible for takeover
+func (n *NpmResponse) NotAvailable() bool {
+	// Check if a known field has a value
+	if len(n.Time.Unpublished.Name) > 0 {
+		return true
+	}
+	return false
+}
+
 // NPMLookup represents a collection of npm packages to be tested for dependency confusion.
 type NPMLookup struct {
 	Packages []string
@@ -88,10 +118,20 @@ func (n *NPMLookup) isAvailableInPublic(pkgname string, retry int) bool {
 		fmt.Printf(" [W] Error when trying to request https://registry.npmjs.org/"+pkgname+"/ : %s\n", err)
 		return false
 	}
+	defer resp.Body.Close()
 	if n.Verbose {
 		fmt.Printf("%s\n", resp.Status)
 	}
 	if resp.StatusCode == http.StatusOK {
+		npmResp := NpmResponse{}
+		body, _ := ioutil.ReadAll(resp.Body)
+		json.Unmarshal(body, &npmResp)
+		if npmResp.NotAvailable() {
+			if n.Verbose {
+				fmt.Printf("[W] Package %s was found, but all its versions are unpublished, making anyone able to takeover the namespace.\n", pkgname)
+			}
+			return false
+		}
 		return true
 	} else if resp.StatusCode == 429 {
 		fmt.Printf(" [!] Server responded with 429 (Too many requests), throttling and retrying...\n")