+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - main
     - New
         - npm: In case package was found, also check if all the package versions have been unpublished. This makes the package vulnerable to takeover
+        - npm: Check for http & https and GitHub version references
     - Changed
 
 - v0.3